[v0.7.5-beta] Tp permissions

+++ routes/agents.rb
@@ -5,15 +5,18 @@ get '/agents/list' do
 end
 
 get '/agents/create' do
+  authorize :application, :admin_access?
   haml :agent_edit
 end
 
 get '/agents/:id/edit' do
+  authorize :application, :admin_access?
   @agent = Agents.first(id: params[:id])
   haml :agent_edit
 end
 
 post '/agents/:id/edit' do
+  authorize :application, :admin_access?
   agent = Agents.first(id: params[:id])
   agent.name = params['name']
   agent.save
@@ -21,12 +24,14 @@ post '/agents/:id/edit' do
 end
 
 get '/agents/:id/delete' do
+  authorize :application, :admin_access?
   agent = Agents.first(id: params[:id])
   agent.destroy
   redirect to('/agents/list')
 end
 
 get '/agents/:id/authorize' do
+  authorize :application, :admin_access?
   agent = Agents.first(id: params[:id])
   agent.status = 'Authorized'
   agent.save
@@ -34,6 +39,7 @@ get '/agents/:id/authorize' do
 end
 
 get '/agents/:id/deauthorize' do
+  authorize :application, :admin_access?
   agent = Agents.first(id: params[:id])
   if agent.status == 'Working'
     flash[:warning] = 'Agent was working. The active task was not stopped and you will not receive the results.'
